vm: post-move fixes
