+++ crates/vm/Cargo.toml
@@ -12,4 +12,58 @@ readme.workspace = true
 repository.workspace = true
 version.workspace = true
 
+[features]
+default = ["wasm-runtime"]
+wasm-runtime = [
+  "parity-wasm",
+  "rayon",
+  "wasm-instrument",
+  "wasmer-cache",
+  "wasmer-compiler-singlepass",
+  "wasmer-vm",
+  "wasmer",
+]
+testing = [
+  "namada_core/testing",
+  "tempfile",
+]
+
 [dependencies]
+namada_core = { path = "../core" }
+namada_events = { path = "../events", default-features = false }
+namada_gas = { path = "../gas" }
+namada_parameters = { path = "../parameters" }
+namada_sdk = { path = "../sdk" }
+namada_state = { path = "../state" }
+namada_token = { path = "../token" }
+namada_tx = { path = "../tx" }
+namada_vp = { path = "../vp" }
+
+borsh.workspace = true
+clru.workspace = true
+parity-wasm = { version = "0.45.0", features = ["sign_ext"], optional = true }
+rayon = { version = "=1.5.3", optional = true }
+smooth-operator.workspace = true
+tempfile = { version = "3.2.0", optional = true }
+thiserror.workspace = true
+tracing.workspace = true
+wasm-instrument = { workspace = true, optional = true }
+wasmer = { workspace = true, optional = true }
+wasmer-cache = { workspace = true, optional = true }
+wasmer-compiler-singlepass = { workspace = true, optional = true }
+wasmer-vm = { workspace = true, optional = true }
+wasmparser.workspace = true
+
+[dev-dependencies]
+namada_core = { path = "../core", default-features = false, features = [
+  "testing",
+] }
+namada_state = { path = "../state", features = ["testing"] }
+namada_test_utils = { path = "../test_utils" }
+namada_tx = { path = "../tx", features = ["testing"] }
+
+assert_matches.workspace = true
+byte-unit.workspace = true
+itertools.workspace = true
+tempfile.workspace = true
+test-log.workspace = true
