+++ crates/vm/src/wasm/host_env.rs
@@ -6,9 +6,9 @@
 use namada_state::{DBIter, StorageHasher, DB};
 use wasmer::{Function, FunctionEnv, Imports};
 
-use crate::vm::host_env::{TxVmEnv, VpEvaluator, VpVmEnv};
-use crate::vm::wasm::memory::WasmMemory;
-use crate::vm::{host_env, WasmCacheAccess};
+use crate::host_env::{TxVmEnv, VpEvaluator, VpVmEnv};
+use crate::wasm::memory::WasmMemory;
+use crate::{host_env, WasmCacheAccess};
 
 /// Prepare imports (memory and host functions) exposed to the vm guest running
 /// transaction code
@@ -116,9 +116,9 @@ mod wrap_tx {
     use namada_state::{DBIter, StorageHasher, DB};
     use wasmer::FunctionEnvMut;
 
-    use crate::vm::host_env::TxVmEnv;
-    use crate::vm::wasm::memory::WasmMemory;
-    use crate::vm::WasmCacheAccess;
+    use crate::host_env::TxVmEnv;
+    use crate::wasm::memory::WasmMemory;
+    use crate::WasmCacheAccess;
 
     pub(super) fn _0<F, RET, D, H, CA>(
         f: F,
@@ -297,9 +297,9 @@ mod wrap_vp {
     use namada_state::{DBIter, StorageHasher, DB};
     use wasmer::FunctionEnvMut;
 
-    use crate::vm::host_env::{VpEvaluator, VpVmEnv};
-    use crate::vm::wasm::memory::WasmMemory;
-    use crate::vm::WasmCacheAccess;
+    use crate::host_env::{VpEvaluator, VpVmEnv};
+    use crate::wasm::memory::WasmMemory;
+    use crate::WasmCacheAccess;
 
     pub(super) fn _0<F, RET, D, H, EVAL, CA>(
         f: F,
